# packages/api-contract/specs/swagger-app.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter by user role",
            "schema": {
              "type": "string",
              "enum": [
                "USER",
                "MODERATOR",
                "ADMIN"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by user status",
            "schema": {
              "type": "string",
              "enum": [
                "ACTIVE",
                "INACTIVE",
                "SUSPENDED",
                "PENDING_VERIFICATION"
              ]
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active state",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "emailVerified",
            "required": false,
            "in": "query",
            "description": "Filter by email verification state",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "fromDate",
            "required": false,
            "in": "query",
            "description": "Filter users from this created date (ISO)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toDate",
            "required": false,
            "in": "query",
            "description": "Filter users to this created date (ISO)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "Search by email, firstName, lastName, phone",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List users (paginated)",
        "tags": [
          "users"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/users/search": {
      "get": {
        "operationId": "UsersController_search",
        "parameters": [
          {
            "name": "query",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Search completed successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search users",
        "tags": [
          "users"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user's profile",
        "tags": [
          "users"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by id",
        "tags": [
          "users"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "users"
        ],
        "x-scope": [
          "app",
          "user"
        ]
      }
    },
...
  },
  "info": {
    "title": "PP02 API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token (Bearer <your_token_here>)"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name",
            "minLength": 2
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "minLength": 2
          },
          "phone": {
            "type": "string",
            "description": "Phone number"
          },
          "avatar": {
            "type": "string",
            "description": "Avatar URL"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED",
              "PENDING_VERIFICATION"
            ],
            "description": "User status"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the user is active"
          },
          "deletedAt": {
            "type": "object",
            "description": "Deleted at timestamp (soft delete)"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}